Master
feat(attachment): unify multimodal asset ingestion engine and modernize workspace UI

Consolidate all developmental milestones from v0.2.7-4 through v0.2.7-7 into a single, production-
ready attachment architecture. This integrates a decoupled polymorphic text parsing framework, 
hardens context safety limits, delivers selective asset removal options, resolves interactive 
UI rendering flaws, and cleans up codebase expressions across the system.

STORAGE & PARSING SUBSYSTEM CHANGES (v0.2.7-4..v0.2.7-5):
- Introduce abstract polymorphic interface ITextParser handled dynamically via ParserFactory to 
  isolate extension-specific formatting logic (.pdf, .docx, .epub, .xlsx, .ods, .odt, .txt, .md).
- Implement an ArchiveReader core using libarchive for container unpacking, implementing absolute 
  bounds checks (m_max_entries, m_max_entry_bytes) to shield memory from Zip-Bomb vectors.
- Protect resource management lifecycles with an explicit Custom ArchiveDeleter functor mapping 
  C-style pointers inside secure std::unique_ptr layouts.
- Stabilize rolling release compilation by mapping pugixml queries to explicit pugi::xpath_node 
  wrapper lists and evaluating poppler::byte_array objects through iterator string range copies.
- Resolve transitive dependency linker faults by explicitly mapping path constraints for 
  ${JPEG_LIBRARIES} and ${PNG_LIBRARIES} directly inside the CMake build layout.
- Fix identifier variable shadowing recursive loops within the Boost.GIL image dimension system.

MULTIMODAL CONTEXT MANAGEMENT & BUFFERING (v0.2.7-6):
- Split metric validation branches to distinguish textual scripts from multi-modal imagery.
- Replace file size calculations for vision files with a unified visual embedding projection 
  allocation layer (1024 tokens) to correspond directly with local hardware vision configurations.
- Enforce strict allocation limitations capping concurrent prompt queues at max_attachment_count 
  (6UZ) and implement specific error paths for IngestionError::MAX_LIMIT_REACHED.
- Expose index-mapped queue optimization logic via AttachmentManager::RemoveByIndex.

UI STABILIZATION, INTERACTION LAYER, & LOGIC SANITIZATION (v0.2.7-7):
- Incorporate active micro-buttons (❌) into individual asset chips, binding inline closure 
  lambdas to remove documents from the attachment manager queue cleanly by numeric index.
- Eradicate the double-rendering chat box bug in append_user_message by verifying incoming 
  prompt duplicate records using string subset lookaheads via .starts_with() == true.
- Expand maximum allowed PDF boundary limits to 4MB, mapping secondary notification structures 
  to address IngestionError::DOCUMENT_TOO_LARGE alerts gracefully.
- Perform a comprehensive codebase sanitation pass, discarding all logical negation operator (!) 
  usages in favor of readable explicit expressions (== false, == nullptr).
- Inject clean textual encapsulation barriers ([START OF CONTEXT DOCUMENT: <filename>]) to cleanly 
  demarcate file data blocks from model prompt instructions during local Ollama inference cycles.

+++ src/engine/storage/history_manager.cpp
@@ -36,7 +36,8 @@ auto HistoryManager::GenerateUuidString() -> std::string {
 
 auto HistoryManager::GetCurrentEpoch() -> uint64_t {
     auto time_point = std::chrono::system_clock::now().time_since_epoch();
-    return std::chrono::duration_cast<std::chrono::seconds>(time_point).count();
+    return static_cast<std::uint64_t>(
+        std::chrono::duration_cast<std::chrono::seconds>(time_point).count());
 }
 
 auto HistoryManager::InitializeDatabase() noexcept -> bool {
